requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

# ML / NLP
torch
numpy
pandas

# Utilities
scikit-learn
